Committing details for graph creation w/Es.

+++ simulator/src/main/java/kcl/teamIndexZero/traffic/simulator/data/features/Lane.java
@@ -39,7 +39,7 @@ public class Lane extends Feature {
      *
      * @return Length in meters
      */
-    public int getLength() {
+    public double getLength() {
         return this.roadSpecs.length;
     }
